Use unsigned long long instead of unsigned long in examples

+++ odb-examples/container/driver.cxx
@@ -52,7 +52,7 @@ main (int argc, char* argv[])
   {
     unique_ptr<database> db (create_database (argc, argv));
 
-    unsigned long id;
+    unsigned long long id;
 
     // Create a persistent person object.
     //
