+++ odb-examples/container/person.hxx
@@ -86,7 +86,7 @@ private:
   person () {}
 
   #pragma db id auto
-  unsigned long id_;
+  unsigned long long id_;
 
   std::string first_;
   std::string last_;
