+++ odb-examples/inverse/employee.hxx
@@ -113,7 +113,7 @@ private:
   position () {}
 
   #pragma db id auto
-  unsigned long id_;
+  unsigned long long id_;
 
   std::string title_;
 
@@ -245,7 +245,7 @@ private:
   employee () {}
 
   #pragma db id auto
-  unsigned long id_;
+  unsigned long long id_;
 
   std::string first_;
   std::string last_;
