+++ odb-examples/relationship/employee.hxx
@@ -139,7 +139,7 @@ private:
   employee () {}
 
   #pragma db id auto
-  unsigned long id_;
+  unsigned long long id_;
 
   std::string first_;
   std::string last_;
