+++ odb-examples/schema/embedded/person.hxx
@@ -49,7 +49,7 @@ private:
   person () {}
 
   #pragma db id auto
-  unsigned long id_;
+  unsigned long long id_;
 
   std::string first_;
   std::string last_;
