+++ odb-examples/section/driver.cxx
@@ -22,7 +22,7 @@ main (int argc, char* argv[])
   {
     unique_ptr<database> db (create_database (argc, argv));
 
-    unsigned long id;
+    unsigned long long id;
 
     // Create a persistent person object.
     //
